🔄 Refactor: 파일 구조 변경 및 svg 이미지 삭제
