+++ package.json
@@ -11,7 +11,7 @@
   },
   "dependencies": {
     "axios": "^1.7.9",
-    "lucide-react": "^0.475.0",
+    "lucide-react": "^0.476.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-quill": "^2.0.0",
