+++ src/stores/modalStore.ts
@@ -16,7 +16,7 @@ interface ModalState {
 }
 
 export const useModalStore = create<ModalState>((set) => ({
-  isOpen: false,
+  isOpen: true,
   text: "",
   cancelText: "취소하기",
   confirmText: "삭제하기",
